# Optimize distance calculations