--- a/pyproject.toml
+++ b/pyproject.toml
@@ -47,7 +47,6 @@ dependencies = [
   "matplotlib>=3.3",
   "numpy>=1.20",
   "pandas>=1.2",
-  "scipy>=1.10.0",
   "tqdm>=4.64.1",
 ]
 optional-dependencies.dev = [