--- a/src/data_morph/shapes/bases/shape.py
+++ b/src/data_morph/shapes/bases/shape.py
@@ -67,7 +67,7 @@ class Shape(ABC):
 
         See Also
         --------
-        scipy.spatial.distance.euclidean : Euclidean distance calculation.
+        numpy.linalg.norm : Euclidean distance calculation.
         """
         return np.linalg.norm(a - b)