<!-- Which issue does this address? -->
Fixes #202

**Describe your changes**

- use `numpy.linalg.norm` instead of `scipy.spatial.distance` (see [this SO comment](https://stackoverflow.com/questions/1401712/how-can-the-euclidean-distance-be-calculated-with-numpy#comment51077024_21986532) and [this SO answer](https://stackoverflow.com/a/47775357) for performance benchmarks)
- cache all `Circle` instances when we create an instance of `Rings`, and compute all separations via numpy

I've benchmarked the above 2 on the `bullseye` and `rings` shapes, and the results are nice:

```plaintext
# bullseye on main
218292773 function calls (215844061 primitive calls) in 85.438 seconds
# bullseye w/ this PR
211510859 function calls (209069591 primitive calls) in 82.753 seconds
# rings on main
224466386 function calls (222017681 primitive calls) in 86.428 seconds
# rings w/ this PR
210318700 function calls (207877434 primitive calls) in 82.098 seconds
```

While the relative performance improvement appears marginal, it only appears that way due to Pandas-related operations being the bottleneck, and for `bullseye` we get **6.7M less** function calls, while for `rings` we get **14.1M less** function calls in total.

Note that no changes were made to the public API.

<!-- TODO -->

**Checklist**

<!-- Place an X between the [ ] for completed tasks -->

- [x] Test cases have been modified/added to cover any code changes.
- [x] Docstrings have been modified/created for any code changes.
- [x] All linting and formatting checks pass (see the [contributing guidelines](https://github.com/stefmolin/data-morph/blob/main/CONTRIBUTING.md) for more information).